requirements.txt
loguru==0.7.2
schedule==1.2.1
pyOpenSSL==24.1.0
pyshark==0.6
pyjks==20.0.0